+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.0.0] - 2023-07-31
+### Removed
+- Removed enum class `DefinedStreams`.
+
+### Changed
+- The strim configuration has been moved to the `config.yaml` file.
+
 ## [0.1.0] - 2023-07-28
 ### Added
 - First version project
@@ -13,4 +20,4 @@ All notable changes to this project will be documented in this file.
 
 [Unreleased]: https://github.com/araucaria-project/oca_nats_config
 
-[0.1.0]: https://github.com/araucaria-project/oca_nats_config
+[1.0.0]: https://github.com/araucaria-project/oca_nats_config